- refactor : 게시판 컴포넌트 저장된 폴더명 변경
