+++ StudentSoup/src/main/frontend/src/pages/board/Board.tsx
@@ -4,8 +4,8 @@ import BoardSearch from '../../components/common/BoardSearch';
 import Paginate from '../../components/common/Paginate';
 import { Desktop, Mobile } from '../../mediaQuery';
 import { getDepartmentBoards, postBoards } from '../../apis/api/BoardAPI';
-import PCBoard from './boardcomponents/PcBoard';
-import MobileBoard from './boardcomponents/MobileBoard';
+import PCBoard from './components/PcBoard';
+import MobileBoard from './components/MobileBoard';
 import { useLocation, useNavigate } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import { type BoardDepartmentType, type BoardDataType } from '../../interfaces/BoardTypes';
