+++ StudentSoup/src/main/frontend/src/pages/board/components/MobileBoard.tsx
