+++ StudentSoup/src/main/frontend/src/pages/board/components/PcBoard.tsx
